moved login to its own function
