+++ src/router.js
@@ -1,7 +1,10 @@
 // imports
 const register = require("./routes/register");
+const login = require("./routes/login")
+
 
 // exports
 module.exports = {
-    register
+    register,
+    login
 }
